Improved performance
Greatly (x20) improved performance by discovering
what the slowest part of the code was, and then trying
to improve that specific part.

I changed, for example, how the movement code checks if a
tile it wants to move to is being occupied by another
creature.
At first, I just had a list with all creatures and for each
movement it went through the entire list checking if one
creature had the same position.
In computer science you can write the complexity of an
algorithm with the "Big O" notation.,
What I had would have been O(n^2) in that notation.
"n" stands (in this case) for the number of creatures on
the board, and as n grows linearly, the time the algorithm
takes goes up exponentially instead!
This blog post describes it better than I did:
https://www.daveperrett.com/articles/2010/12/07/comp-sci-101-big-o-notation/

I rewrote the algorithm until it had a time complexity
of O(N), which means that the time the algorithm takes now
scales linearly with the amount of creatures!!!